Encounter type field in encounter form.  Fixes #5260 (#5272)
* WIP encounter type field in encounter form.

Since we can't bundle certain codes like CPT4 or even SNOMED in
non-member countries we want to only install the codes property for
the Encounter Type list options if we have the right code systems
installed.  So we have added a CodeTypes event subscriber that listens
to newly installed coded types as well as the SQL Upgrade script process
to update the Encounter Type codes property.

This is WIP and is not finalized.

* Fixes #5260 Encounter type field

Changed encounter form to save the encounter_type_code and encounter_type_description.
If we have CPT4/SNOMED-CT codes installed we will use those for the code and description.
If we don't have the codes installed we will just use the list option option_id and title for the code
and description.  Not valid for CQM but it allows other non-US based systems to continue to function.

Implemented a CodesType module that allows us to hook into the upgrade and code installation process
and do any module processing such as mapping our list option codes to their appropriate code sets.

Added the encounter-types to the database file for new installations.

Removed default column on first encounter-types list-option.