Immunizations UI add reason_code support
Added the reason_code and reason_description to the Immunizations UI so
reasons can be specified as to why the immunization was done.  Required
for Clinical Quality Measures but useful for billing purposes as well.
If there is a refusal reason there is a dropdown to select from the
supported value set of possible refusal reasons.