Fixes #5276 Procedure Order Reason Codes
Added to the procedure order UI the reason code, reason status,
reason description, start and end dates for the reasons. Now users can
hit the reason code icon and it will expand to provide the additional
reason code, status, start/end dates.

Refactored the addProcLine to use a template from the HTML dom so we can
also handle the reason code expanded widget. Tested the various ways
addProcLine to make sure it still works in the various ways it is
called.

Changed the Procedure Order Date Ordered to be a datetime selector
instead of just a date selector.