Openemr fix 5277 immunization reason codes (#5296)
* WIP #5277 Immunization reason codes

This depends on the encounter_type work so we are making this a WIP for
now.  We need to handle both the reason and negationReason which for
immunization is the refusal_reason code.  I couldn't find any examples
where both a reason and negationRationale QDM exist simultaneously but
the spec supports the use case.  Since we already had reason codes
populated I map the valuesets for QDM to the NIP002 CDC code mappings in
the SQL so we can still support reporting to immunization registries.

I also updated the FHIR to support the religious objection since that's
the only additional value code that makes sense in addition to patient
objection for what FHIR allows.  I imagine future versions of FHIR will
be more in line with CCDA/QRDA for the immunization reason codes.

* Immunization populate code if snomeds installed

If the OpenEMR installation has the SNOMED code types installed we will
go through and populate our code types so that we can make them
available for use.

Fixed the sql upgrade script.

* Immunizations UI add reason_code support

Added the reason_code and reason_description to the Immunizations UI so
reasons can be specified as to why the immunization was done.  Required
for Clinical Quality Measures but useful for billing purposes as well.
If there is a refusal reason there is a dropdown to select from the
supported value set of possible refusal reasons.

* Style fixes

* Added immunization data to database.sql

* Fix text escaping and db comments