WIP FHIR $docref operation #5284
Added the infrastructure for the $docref operation.  Added scopes, rest
endpoint, and service classes.

Currently the endpoints just return an empty DocumentReference.  Need to build
it to return the DocumentReference for the generated ccda.

Moved the Export operation into a sub folder where we can keep our future operations that we support.