(g)(9) ccda gen, (b)(1) sort/truncate ccda
Implemented generating a ccda via an event in the system so fhir can create
the ccda for the $doc-ref FHIR operation.

Refactored the ccda generation pieces from the Encounterccdadispatchcontroller
into separate classes so we can call it from the event cleanly without having
to deal with a ton of POST / controller logic code.

Refactored the ccda document generation so that it uses the document class for
encryption, couchdb storage, and retrieval.  This means a generated ccda now
generates a file in the patient CCD/CCDA folder.

Implemented a view ccda event that allows the Carecoordination module to sort
and truncate the ccda sections for the ccda that is being viewed.  This is
necessary in order to meet the b1 requirements of a user being able to set
how many sections they want to see as well as the sort order of the sections.

Began the work of adding a new globals field for displaying a list of options
and adding that list to the globals.  The options are saved in the globals
table as a semi-colon delimited string.

Added the beginnigs of the ccda-sections list options that people can select
from in the globals user settings table.

Modified the FHIR Document Binary api to use uuids instead of ids for the
document. Made the ids backwards compatable so old Documents using the id columns will
still work.

Added a method to the CodesTypeService to parse codes into codeable concepts
so we can use them in FHIR.

Updated the CCR/display.php file to use the user preferences and ccda view
event so that we display CCDs with the newest stylesheet that @sjpadgett
implemented.

Added to the CCR/display.php checks to make sure the document exists and that
the ACLs are being properly checked.

Added documentation to some of the Carecoordination modules.

Added to the GlobalsService the addUserSpecificTab method in order to be able
to add additional tabs to the user preferences section.