CCDA zip/saved file filename changes
Changed up the CCDA generation filenames so we have date timestamps and the
type of ccda that is being genreated in the filename and the zip.
Changed up the direct email attachment filename to include the ccda filename
if one is present so we have the exact file sent that we downloaded.

Changed up the CCDA result object to have the filename that was generated for
the file.

Refactored the sendTo.js::send() method to break up the download_all and
direct send into their own functions that can be better reasoned about and
that transmit the document type needed to generate the xml / zip filename.

Changed out the Document filename for the generated ccda to being a human
readable name and not just a uuid.

Fixed a bug in the amc logging that would fail if a ccda is sent w/o there
being a created referral record first.

Modified Direct transmission to include a filename if one is available