Openemr fix 5284 fhir docref (#5325)
* WIP FHIR $docref operation #5284

Added the infrastructure for the $docref operation.  Added scopes, rest
endpoint, and service classes.

Currently the endpoints just return an empty DocumentReference.  Need to build
it to return the DocumentReference for the generated ccda.

Moved the Export operation into a sub folder where we can keep our future operations that we support.

* (g)(9) ccda gen, (b)(1) sort/truncate ccda

Implemented generating a ccda via an event in the system so fhir can create
the ccda for the $doc-ref FHIR operation.

Refactored the ccda generation pieces from the Encounterccdadispatchcontroller
into separate classes so we can call it from the event cleanly without having
to deal with a ton of POST / controller logic code.

Refactored the ccda document generation so that it uses the document class for
encryption, couchdb storage, and retrieval.  This means a generated ccda now
generates a file in the patient CCD/CCDA folder.

Implemented a view ccda event that allows the Carecoordination module to sort
and truncate the ccda sections for the ccda that is being viewed.  This is
necessary in order to meet the b1 requirements of a user being able to set
how many sections they want to see as well as the sort order of the sections.

Began the work of adding a new globals field for displaying a list of options
and adding that list to the globals.  The options are saved in the globals
table as a semi-colon delimited string.

Added the beginnigs of the ccda-sections list options that people can select
from in the globals user settings table.

Modified the FHIR Document Binary api to use uuids instead of ids for the
document. Made the ids backwards compatable so old Documents using the id columns will
still work.

Added a method to the CodesTypeService to parse codes into codeable concepts
so we can use them in FHIR.

Updated the CCR/display.php file to use the user preferences and ccda view
event so that we display CCDs with the newest stylesheet that @sjpadgett
implemented.

Added to the CCR/display.php checks to make sure the document exists and that
the ACLs are being properly checked.

Added documentation to some of the Carecoordination modules.

Added to the GlobalsService the addUserSpecificTab method in order to be able
to add additional tabs to the user preferences section.

* CCDA improvements, new global settings list type

Modified ccda node service to skip observation interpretations that are empty
for lab results.

Modified the encounter ccda controller to catch ccda node service errors and
message the user to check their ccda node service installation instructions.

Implemented a ccda sort feature needed for (b)(1) using the user preferences
globals.

Updated ccda to support standard units (inches,lbs) with the right unit codes.

Cleaned up the ccda createDocument per code review suggestions.

Added the capability to the globals to have a list+add option that is sortable.

Made it possible to have additional field options for global settings.

Added the ccda-sections list option that has all of the sortable sections we
implement in CCDA.

* Style fixes

* CCDA zip/saved file filename changes

Changed up the CCDA generation filenames so we have date timestamps and the
type of ccda that is being genreated in the filename and the zip.
Changed up the direct email attachment filename to include the ccda filename
if one is present so we have the exact file sent that we downloaded.

Changed up the CCDA result object to have the filename that was generated for
the file.

Refactored the sendTo.js::send() method to break up the download_all and
direct send into their own functions that can be better reasoned about and
that transmit the document type needed to generate the xml / zip filename.

Changed out the Document filename for the generated ccda to being a human
readable name and not just a uuid.

Fixed a bug in the amc logging that would fail if a ccda is sent w/o there
being a created referral record first.

Modified Direct transmission to include a filename if one is available

* style fixes

* Fix db upgrade extra sql