Multiple Patient Address Initial Support
Added a new data type for LBF to support Address List data type.

Addresses can be added with multiple addresses supported.  Initial
framework for saving addresses is here as well s retrieving them.

Rolled out the initial contact and contact address classes that can
be used to support Patients.  Decided to leverage the ORM classes
for Contact and ContactAddress to quickly persist and model
the data.