Fixes #5450 viewer display options
Implemented the display all / hide all functionality.  Moved to an
in-memory state for the sections instead of using the DOM to track
visible / invisible.

Also added a message if all the sections are hidden to make the
presentation look nicer.

Fixed some minor bugs on the hide/display.

Got everything working with the section re-arranging so people can
change up the order of the sections now if they want to print a
different display order than what they have in their preferences.