Openemr fix 5435 direct message validation report (#5439)
* WIP #5435 Display Validation Report

Moved the linked document messages into its own file just for sanity
sake.  Eventually we can move this to twig when we can migrate the whole
file.

Wrote a helper object for validating the CDA for a given Document class.
We need to do some more introspection in order to make sure we only
validate CDA documents and not just all zips and all xml files.  We need
some additional formatting on the validation report pieces.

* Fix style issues

* Refactored ccda validator messages display.

Fixed some minor typos and UX display of the ccda documents.

Refactored the carecoordination module and the messages to use the same
html template for the error display

* Fix styles

* Fix javascript bugs in linked documents

* Fixes #5435 Connect async ccda validation process

Make the ccda validation an asyncronous process.  Give the option of
calling the async process as a json file for future developers.

Provide error count numbers as part of the validation.  Handle multiple
file validations.

Skip validation of non-cda files.

* Remove public mdht validation service

We don't want people accidently submitting PHI to a government public
server so we are removing the default government option.  We change the
description so people can know how to deploy their own CDA validation
service.

* Fix code review escaping,comments.

* Fix php namespace error

* Fix the php errors with js_url escaping.

* Add in exit if script is called directly.

* Fix escaping issues, document contents error

Went through and manually triggered all the json files to make sure they
are working properly.  Hopefully we've resolved all the issues at this
point.

We handle the JSON files as I wanted to make sure people could validate
documents using the JSON syntax if they need document validation and
need to do their own rendering.  Also went through the html files as
well with escaping.