Fixes #5455 Document Type Display Order
We decided to implement this by having multiple sort order sections for
the four document types we support in CCD-A.

I wrote a new preview option for CCD-A that examines the XML document to
figure out the correct CCDA Template document type so we can know what sort order to use to
display for the correct document.

Added a deprecation notice to the ccr/display.php in case anyone is
using it.  We no longer reference it from the documents pages so we can
probably remove it at some point.

I also added a new filter event to the patient document tree viewer so
people can modify / update the html tree rendering on the documents if
they so choose.  This made it easy to loosely couple the tree with the
Carecoordination module.