Fixes #5473 Metric Precision fixes
Increased the decimal precision from 2 digit to 6 digits in the database
for form_vitals.  Our metric conversion was not giving back the same
values to the user as what they were entering in.  Much of this was due
to the conversion precision of just 2 digits.

Also MySQL is deprecating the float(X,N) data type and so we changed to
the decimal format to make this work.  Note that upgrades will need to
be done in off hours here as we are dropping the database columns to
handle the conversion.

Changed up the vitals conversion javascript so we could have a hidden
input that contains the 6 digit precision for metric / us conversion
purposes but only shows the two digit value to the user to keep the
interface clean and less confusing.

precision to users