Fix vitals javascript errors.
kg to lb was throwing an error as we tried calling a Number method on a
string.

Vitals were saving in metric if you had metric set as the main unit and
entered in data that way.

Setting a vitals value to be empty was not emptying out the
corresponding conversion field and the saved input.