Address widget date selectors,county,reuse lbf
Made it so the address widget re-uses the lbf widget display code for
each of the address element codes.

Refactored the tables into divs to be mobile friendly.

Added in the date start and end dates so we can support previous
addresses.

Added the address use and address type selectors so people can identify
what kind of address it is.

Exposed the option js functions into the global window under the oeUI
container so we can have our add button on the state, country, use, and
type selector widgets.