Fixes #5367 previous address support
Implemented the saving of the address list lbf template in the new
patient and the existing patient demographics page.

Added a new lbf field called additional_addresses to be displayed in the
contacts group layout.

Added the district field to support counties / districts in addresses.
Added the use and type FHIR data fields.
Also added a period start and period end to the addresses so we can
track the time period an address was in use.  New addresses default to a
current date of today with an empty end date.

Added new list options for address-uses and address-types to support the
new FHIR options.  The option_id is the FHIR key name.

Migrated some of the date formatting into the DateFormatterUtils object
so we can leverage them inside our objects without having to include the
formatter.inc.php file.

Added the address_list datatype to the restricted lbf fields so users
can't remove the datatype since we have to have it for USCDI support.

Refactored all of the styles of the address list widget to be responsive
and mobile friendly using bootstrap classes and styles.

Made the js functions of addresses consistent in naming conventions with
our other js functions.

Implemented the archiving / deactivating of the address list items.  Had
to switch from the BINARY(1) data type to a char value.  The
ORDataObject has a bug in that you can't save a value with a numeric 0
as its treated as empty.  I didn't want to try and change the logic and
mess up the other classes that extend ORDataObject, so in the interest
of time I switched isPrimary and status to be char fields that we key
off of.

Renamed the contact table fields to be consistent with fire naming
conventions per @mdsupport's recommendation.

Changed up the js events to be added in javacsript and not using inline
onclick events.  This simplifies the debugging immensely.