Initial implementation of #5508 ccd date filtering
Changes g9 $docref to be a POST api since we generate a new document
every time.

Fixes ISO8601 date problems with php since php's date constant is
incorrect.  Switch to ATOM constant.

Got the start and end date working from the g9 but we are failing in
some instances in the ccda generator because its excluding encounters
that some of the data elements rely upon.

Fixed the ORDataObject saving incorrectly on some database instance
types due to the timezone and microsecond formatting.

Added __clone helper methods to the search field classes so we can copy
them and modify properties w/o modifying the references.