Fix detailed race display
Had a case where the race code display was 'White' and the detailed race
display was 'White European'.   This made it look really odd and funny.
Put the detailed race code on its own line.  Found out its a 0.*
relationship so made it so we can have multiple lines depending on how
many race codes there are.