Fixes #5543 patient demographic display in xsl (#5544)
* Fixes #5543 patient demographic display in xsl

Made patient display name be the Legal name if available
Added display of additional patient names in header
Added display of sdtcRace and sdtcEthnicity values
Added multiple address display, multiple telecom display
Added support for Mobile Contact telecom

* Fixes #5546 preview ccda button

Added a preview button to see the ccda right away without having to go
to the documents screen and categorize the document as a ccda.

* Fix detailed race display

Had a case where the race code display was 'White' and the detailed race
display was 'White European'.   This made it look really odd and funny.
Put the detailed race code on its own line.  Found out its a 0.*
relationship so made it so we can have multiple lines depending on how
many race codes there are.

* Fixes #5545 unknown doc sort order

Fixed an issue with the document sort order breaking if an unknown
document type was discovered.  Also fixed an error if the document had
comments and we were trimming the document.  Had to use elements instead
of node children.

* Fix birthname errors in cda stylesheet

If the given name was a birthname record but the family name is not
marked as birthname we were not displaying the birth name label.

Fixed it if an unknown name type is found that we show the Previous Name
label.

* Add performer and performer role to cda stylesheet