ccda author and dates (#5547)
* ccda author and dates
- 3 chances for author
- datetime validation changes

* - fix using the wrong template header author tele +1

* Fix ccda author,custodian,config settings (#21)

Made it so the author npi works and that the author doesn't blow up when
it returns as a string.

Fixed issue with the custodian uuid accidently being set as the phone
number.

Added in the author date time stamp override in the config settings.

* - add primary diagnosis to documentationOf template and CCM

* Ccda author encounter provenance flag (#22)

* Fix ccda author,custodian,config settings

Made it so the author npi works and that the author doesn't blow up when
it returns as a string.

Fixed issue with the custodian uuid accidently being set as the phone
number.

Added in the author date time stamp override in the config settings.

* Callback participant, force encounter date flag

Added an office contact setting in care coordination.  Added the
participants piece to the header.

Removed the fDate formatting on the document timestamp as PHP was
already sending a timestamp formatted document and the fDate was being
corrupted.  Had to do this to pass ETT validation.

Added a checkbox to let people force the author timestamp value to be
the encounter date time stamp.  This made it possible to override the
timestamps for the document across the three different test scenarios
without having to change it for each individual patient.  Most people
will have this flag turned off, but it sped up the ONC testing instead
of having to set the timestamp on each doc generation (for both b1 and
g9 testing).

* Remove extraneous letter in send screen.

* Add referring participant refactor participants

Refactored the participants on the php side to be an array so we can add
participants in the future.  Added the referral participant

* Fix document effectiveTime, participant datetime

Got the effectiveTime back to using moment like Jerry and I discussed.

Got the participant time working since referral date has its own format
regardless of the system format.

* Add telecom to participant

Wasn't passing ETT test cases w/o telecom info.

* Fix g9 filtering problems

Had problems on the single date filtering test.  This commit fixes
those.

* Fix g9 date displays, stylesheet timezone bug

Had a bug where the timezone was showing but no hour/minute because they
get hidden if the time is 00:00.  Made the timezone hidden if the
hour/minute is also hidden.

Fixed the g9 service dates not rendering because they were populated in
a range request.

* Fixed zip and street address for participants.

Got the referring and contact address into the stylesheet.  Definitely
not ideal as the header is huge.  Will have to touch up the styles and
add in the fallback PCP if the REFerral isn't there for our human
readable view.

* - add zime zone to encounter date
- fix some minor participant issues (I think issues for windows only)
- add componentOf template
- add provider code to author
- add utc offset for est for any date with a time zone or tz flag
- refactor all author dates to passed in date from CCM
- fix empty addresses with conditionals
- remove spurious text

Co-authored-by: Stephen Nielson <stephen.nielson@gmail.com>