Fixes #5565 referral fallback, style fixes (#5566)
Implemented the referral fallback by using the encompassing encounter
provider information if we have one in the displayed CCDA, or falling
back to the Primary Care Provider (PP) or the Primary Care Physician
(PCP) if we have one.

In order to grab the phone and contact information I had to fix up some
issues in the care team and provider templates.  Those are still
displaying properly in the documentationOf and in the Care Team
sections.

Fixed up a number of style issues in the cda to make the format a bit
more condensed while still preserving readability.  Did this by moving
address info all into one line.  Found a couple of places where headers
were being pushed onto two lines that I tidied up.