Provenance Authorship for patient data
Add created_by and updated_by columns to patient_history, patient_data,
and history_data as well as modify dates and created dates so we can at
least track the last user who modified a piece of information and the
user that created a piece of information for provenance purposes.

A more detailed history can then be derived from the audit table log as
we'll be able to see what user modified the data on each patient record.