Fixes #5643 ModulesClassLoader
Implemented a modules class loader that allow module writers to register
a namespace if its not already been registered.  This makes it so the
ModuleWriter doesn't have to use requires everywhere and a module can be
just dropped into the filesystem and execute once its been registered
and installed from the Manage Modules interface.