Convert Vitals form from smarty to twig
Big overhaul of vitals from changing it from smarty to twig format.
Cleaned up a lot of the rendering as well so we can both easily extend
the vitals form and handle additional fields.  This is prep work to make
the form responsive.

Added some additional extension to twig so we could handle the wierd
date formats for the vital forms.