Fixes #5681 Header events for js/css
Fixes #5681
Added the ability to inject javascript or css into any page inside
OpenEMR using the Header class specifically the setupHeader method which
is used in smarty, twig, and our regular php files all over the place to
bring in javascript and css assets.  This allows module / event writers
to inject content on every page or on specific pages based on the script
name / page name.