Rel 610 Bring in  #5643 #5681 #5683 #5685 (#5692)
* Fixes #5643 ModulesClassLoader autoload module code without composer install (#5644)

* Fixes #5643 ModulesClassLoader

Implemented a modules class loader that allow module writers to register
a namespace if its not already been registered.  This makes it so the
ModuleWriter doesn't have to use requires everywhere and a module can be
just dropped into the filesystem and execute once its been registered
and installed from the Manage Modules interface.

* Style fixes

* Fixes #5681 Header events for js/css (#5682)

Fixes #5681
Added the ability to inject javascript or css into any page inside
OpenEMR using the Header class specifically the setupHeader method which
is used in smarty, twig, and our regular php files all over the place to
bring in javascript and css assets.  This allows module / event writers
to inject content on every page or on specific pages based on the script
name / page name.

* Fixes #5683 user admin events (#5684)

* Fixes #5683 user admin events

Fixes #5683 adds events to user_admin and usergroup_admin_add page in
order to be able to render content on those pages from a module.

* Remove debug line

* Fixes #5685 add provider calendar methods (#5686)

* Fixes #5685 add provider calendar methods

* Fix spelling, improve sql escaping

* Add QueryUtils to service.