Fixes #5708 Add Appointment FHIR,Portal endpoints (#5709)
* Fixes #5708 Add Appointment FHIR,Portal endpoints

Fixes #5708
Implemented a very basic version of the FHIR R4 Appointment resource.

Opened up the appointment resources for patient standalone apps with the
patient portal if people want to consume direct appointment resources
from the api.

* Add patient pid back to AppointmentService

Because of the way the search() was refactored we need to expose more
columns such as the pc_pid column in the function returns.  Had 3rd
party modules breaking in using the AppointmentService class.

* Move database changes for PR, bump db version

* Add Unique uuid for database.