Fix scope permission on insert/create/delete
Fixes #5834 - Add delete interaction in capability statement to handle
write scopes correctly.

We were using the wrong interaction name for the capability statement
and I renamed it in the prior commit.  However, I wasn't handling it in
our ScopeRepository correctly.  Changed the interaction reference to be
'create' instead of 'insert' and put in a documentation url of where to
go to get the reference.