Fixes #5877 Comlink Telehealth Module
Adds a telehealth module to OpenEMR for 1:1 video teleconferencing.

This module connects to hooks in the calendar, main screen, patient
demographics, user demographics, and patient portal credentials screen
on the OpenEMR provider interface.

On the patient portal this module connects to hooks in the appointment
screen to be able to launch telehealth sessions.

The main launching class is the Bootstrap.php class that connects to the
OpenEMR hooks.

There are several ajax communications that occur such as registering
providers / patients for telehealth services when they login to the
application.  These are all communicated via the public folder.
Index.php handles the provider ajax communications.  index-portal.php
handles all of the AJAX communication for the patient side.

Several tables are created by this module.
- comlink-telehealth_auth which holds the user provisioning /
authentication information for the Comlink backend telehealth servers.
- comlink_telehealth_person_settings holds individual telehealth
settings for providers and patients
- comlink_telehealth_appointment_session - holds a specific telehealth
session information that connects a provider w/ a patient