Fix cache busting for calendar
Cache busting was already occurring in the script filter events via the
Header class and the double cache busting parameter was messing up the
frontend dispatcher.  This fixes that problem.

Also if the URL already has a query param, I fixed the header so it only
appends the param if there isn't a query parameter already.