Fixes #5877 Comlink Telehealth Module (#5878) (#5884)
* Fixes #5877 Comlink Telehealth Module

Adds a telehealth module to OpenEMR for 1:1 video teleconferencing.

This module connects to hooks in the calendar, main screen, patient
demographics, user demographics, and patient portal credentials screen
on the OpenEMR provider interface.

On the patient portal this module connects to hooks in the appointment
screen to be able to launch telehealth sessions.

The main launching class is the Bootstrap.php class that connects to the
OpenEMR hooks.

There are several ajax communications that occur such as registering
providers / patients for telehealth services when they login to the
application.  These are all communicated via the public folder.
Index.php handles the provider ajax communications.  index-portal.php
handles all of the AJAX communication for the patient side.

Several tables are created by this module.
- comlink-telehealth_auth which holds the user provisioning /
authentication information for the Comlink backend telehealth servers.
- comlink_telehealth_person_settings holds individual telehealth
settings for providers and patients
- comlink_telehealth_appointment_session - holds a specific telehealth
session information that connects a provider w/ a patient

* Fix black stylesheet property

* Fix telehealth readme with updated instructions

* Add cache busting to js scripts

Adding in caching parameters exposes a problem in the ModulesApplication
where it wasn't handling script url query parameters.

* Fix sql if statement

* Fix spaces in translations, duplicate config value

* Fix translations and escaping

* Fix html and js escaping.

* Add cache buster to calendar css

* Add text escaping to appointment notes

* Fix style changes

* Add space between name and waiting message

* Make portal credentials fit on one page

With extra verbiage on the screen we reduced the font from 24pt to 22pt
so it can all fit on one page.

* Telehealth module Fix list option population

* Fix cache busting for calendar

Cache busting was already occurring in the script filter events via the
Header class and the double cache busting parameter was messing up the
frontend dispatcher.  This fixes that problem.

Also if the URL already has a query param, I fixed the header so it only
appends the param if there isn't a query parameter already.

* Remove extraneous console.logs

* Telehealth fix unit tests, security

Make it so the unit tests can only run from the command line.  Fix
extraneous parameter in unit test.

* Changed up the unit test a bit to function properly

* Remove extraneous escaping text calls

* Fix translation escaping problems

* Fix escaping issues