Fixes #5919 fix timezone code to handle DST/UTC
We aren't handling timezones properly in what we return for read
operations via FHIR.  We've been returning times as UTC as if the time
was in the UTC/GMT timezone which is incorrect.  We've been needing to
report the timezone information as a UTC time WITH the local offset
expressed in the timezone.  Part of this problem is the fact that we
don't store times in UTC internally.

This fixes #5919 hopefully for good.