Fixes #5943 and additional logging
Added to our dispatch configuration that if we have a 'launch' scope
context in addition to the 'launch/patient' context we still populate
the context information which is really what we want anyway as we want
the request to be bound to what was selected in the access token.

Added some more logging to make sure in debug mode we can figure out
what the context is being populated with.