Rel 700 backport #5919 #5931 #5932 #5933 #5940 #5943 (#5956)
* Fix #5919 datetime timezone. (#5920)

Handles the timezone requests on the date search field to parse out the
requestor's timezone.

Handles the adjustment of the requestor's timezone on the search field
statement resolver to be the local time.

* Fixes #5919 fix timezone code to handle DST/UTC (#5929)

We aren't handling timezones properly in what we return for read
operations via FHIR.  We've been returning times as UTC as if the time
was in the UTC/GMT timezone which is incorrect.  We've been needing to
report the timezone information as a UTC time WITH the local offset
expressed in the timezone.  Part of this problem is the fact that we
don't store times in UTC internally.

This fixes #5919 hopefully for good.

* Fix #5931 FHIR Binary Download (#5934)

Fixes #5931 Make the patient uuid optional for provider/system requests.

* Fixes #5932 don't fail on empty group export (#5935)

Makes it so we don't die w/ a 500 when there are no patients in a group
to export.  It just returns on the request.

* Fixes #5933 Fhir Location Type invalid (#5936)

We were using invalid types in the Location.type field which I've gone
ahead and fixed.

* Fixes #5940 convert uuid for surgery recorder (#5942)

This fixes #5940 where the recorder_uuid was failing.

* Openemr fix 5943 patient context launch scope (#5944)

* Fixes #5943 and additional logging

Added to our dispatch configuration that if we have a 'launch' scope
context in addition to the 'launch/patient' context we still populate
the context information which is really what we want anyway as we want
the request to be bound to what was selected in the access token.

Added some more logging to make sure in debug mode we can figure out
what the context is being populated with.

* Style fixes