Initial screensharing commit WIP
Broke up the telehealth.js into separate files to make it easier to
reason and compose the pieces together.  Initial work on the
screensharing pieces and updated the comlink bridge library to support
the screensharing.

Added exception handling for when the bridge is down to notify the user
of the problem.

Added a settings verification piece in the module.