WIP patient screensharing, 3rd party setup
Did work to help setup calls and screensharing so we can have more than
one party on the call.

Also got the screensharing working from patient to provider.  Still
having bugs with screensharing from provider to patient which is the
next step to work on.

Updated the cvb version so we can receive the event for the screenshare
end.