WIP implement ability for 3rd party calls.
On the backend we send down a participant list when a user attempts to
join the room of who is in the room, etc so we can know who to make the
calls to.

On the frontend we allow those users who are in the participant list to
join the call.

We still need to handle the third party attempting to join and letting
them in.