WIP Telehealth implement add patient logic
Implemented the backend database structure for third party invitation to
the telehealth session.

Implemented creating a patient from the provider to invite them as a 3rd
party to the telehealth sesison.

Setup initial work for searching for duplicate patients to avoid
re-creating patients.

Changed out the settings configuration on the frontend for the provider
to be an add patient dialog.