Fixes #6190 Portal Login Redirect
This change allows the portal page to redirect to another page other
than the SPA home.php file once a user has authenticated.  In this
initial release we only allow this to go to a module page that exists in the
OpenEMR directory.  As needed we can open this up to other areas of the
application.

The redirect param is retained across password changes and the mfa auth
that is handled in the portal/index.php.

I modified the verify_sessions so that the landingpage param can be
updated so the script can be used inside of other pages we want to be
protected by the portal session (such as inside a module).