Fix third party screen display w/ missing patient
If the patient was not on the call but the third party user had joined
the 3rd party would not show on the main screen.  Changed it so the 3rd
party shows on the main screen, made it so if the patient joins it
refocuses on the patient for the provider.