Fixes #6190 Portal Login Redirect (#6191)
* Fixes #6190 Portal Login Redirect

This change allows the portal page to redirect to another page other
than the SPA home.php file once a user has authenticated.  In this
initial release we only allow this to go to a module page that exists in the
OpenEMR directory.  As needed we can open this up to other areas of the
application.

The redirect param is retained across password changes and the mfa auth
that is handled in the portal/index.php.

I modified the verify_sessions so that the landingpage param can be
updated so the script can be used inside of other pages we want to be
protected by the portal session (such as inside a module).

* Switch to !empty instead of isset

* Portal Login change input type text to be hidden

Change the redirect input type from text to be hidden.

* Fix the encodeUri and handle reset password

Need to pass the redirectUrl for the reset password if the recaptcha is
enabled.

Also fix the encodeUri escaping on the js.

* Fix style problems.