Thirdparty screenshare, invitation dialog changes
Fixed third party screenshare calls as screenshare was working for only one participant.
Added the name of the caller to the title property for the videos.  Now when
you hover the mouse over the video it shows who the caller is.  Helpful for
debugging.

Added the current third party invitation to the dialog screen so we can see
who was invited.

Changed the dialog screen to update the new participant when the screen loads.

Updated the presentation screen css