Telehealth Pinned caller, participant label
Made a participant label show up for remote callers.  Still need a label
for the local caller.

Added in the toggling of a pinned caller to keep the focus on that
caller.  Clicking on the video toggle sets the pinned status.  Need to
handle the user clicking the pinned icon as well to toggle / untoggle.