Telehealth major UX update to support mobile
Changed up the user interface to support mobile better.  Moved the
participant list to be an overlay that can be expanded and contracted.

Made the video control bar icons more responsive, changed them to use a
light on dark contrast.

Refactored a lot of twig files to better support testing.

Made it so the video controls hide on inactivity and show up on mouse
movement or on clicking the main screen on the video.