PSR fixes, telehealth invitation event hooks.
Fixed the PSR issues.

Added a TelehealthNotificationSendEvent class event that other modules
can listen to and hook into when a notification event is being sent.

This allows event listeners to change up the recipients, change up the
message or do additional notification transports.  For example the
notification could be sent via text message as well for the invitation.