Add patient dialog fixes, participant list bug fix
Had a bug where if the patient is in the session and the provider adds
another patient to the call, the third party patient can't join until
the patient leaves the call and joins again.  This happened because the
patient was not receiving the updated participant list.  Now everytime
there is a session update (every 5 seconds) it sends down a basic list of the
participants in the call. If the participants are different than the
local client has, the local client hits the server to get the full
participant details.  This makes it so the third party can now join the
call to the patient when they join the session.  Took a while to find
and fix this bug.

Changed up the patient dialog so it updates the patient list when a
patient succesfully is added.